Added an optional dependency of pyhcl
In order to more easily work with data returned from the `get_policy`
method I added the `pyhcl` parsing library to optionally return the
policy data as a Python dict. This has been added as an optional
dependency so that users of `hvac` are not required to have another
dependency in their projects if they don't need the additional functionality.