refactor(auth/token): Break out token roles functionality
I wanted to get a head start on this while the token roles stuff was
fresh in my mind!

I've added two new classes: ClientFeature and TokenRoles. The former is a
wrapper around the client for the basic API calls (post, read, list and
delete). The later inherits the former and is where the new token roles
functionality lives.

Also added further checks to current tests.

Motivation for changes is discussed in #95.

BREAKING CHANGES:

* token_role(...) -> token_roles.read(...)
* list_token_roles(...) -> token_roles.list(...)
* create_token_role(...) -> token_roles.write(...)
* delete_token_role(...) -> token_roles.delete(...)