Change method of checking value in policies
Vault seems to have changed the way that the API returns policies
for the userpass backend between 0.7.x and 0.9.x. In 0.7.3 the
value is a comma-separated string, and may contain the value
"default". In 0.9.3 it's a list, and doesn't include "default".
Using 'in' should account for either use case, though note that
we aren't dealing with Vault's breakage in backward-compatibility.