Fix premature read on stream requests
When a caller (such as `sys.take_raft_snapshot`) performs a stream
request, the act of attempting to parse the response as JSON causes the
entire response body to be read and the underlying connection to be
closed. This renders the streaming response moot.

This change addresses that issue by examining if the caller requested a
stream response, and if so returning the response as is.

Without the change, it is impossible to read raft snapshots that are
larger than memory as the entire response is read into memory to attempt
to read as JSON.