setup.py: Add encoding for long_description
Without an encoding enabled, open() will default to the platform's default encoding. For *nix, this is fine as utf-8 is the default encoding. For windows, the default is CP-1252. This is a problem if something like an emoji is added to the readme, where open() will fail parsing the unicode character. This breaks installing the wheel for Windows machines.

Specifying the encoding as UTF-8 should ensure the README.md can be parsed by both platforms.

Signed-off-by: Colin McAllister <colinmca242@gmail.com>