Token: Add create orphaned token method
The Vault API allows orphaned tokens to be created without root
privileges using the create-orphan endpoint, a separate method should
exist to use that endpoint. This allows implementers the choice of using
the create endpoint with the no_parent parameter set to true or with the
new create_orphan method.

Signed-off-by: Colin McAllister <colinmca242@gmail.com>