Integration Tests: Fix GCP test error
It appears that after HVAC version 1.10, the error code returned by
attepting to auth to an auth backend mounted at a different path than
the default returns a Forbidden response code and different error
message. The test is now updated to handle the differnt return codes
depending if the version is less than version 1.10.0 or not.

Signed-off-by: Colin McAllister <colinmca242@gmail.com>