Tests: Fix Github auth tests
The Github authentication tests broke after hvac made changes in version
1.10.0 where organization ID is used instead of name. This modifies
Vault's behavior to query Github for the organization ID when the Github
auth method is configured. In addition, one of the failure responses
changed from an invalid request to an internal server error.

The Github mock request handler has been updated to provide an org
endpoint for Vault to use. The modified response has now been
incorporated as well.

Signed-off-by: Colin McAllister <colinmca242@gmail.com>