Cert: Fix role certificate parameter
The `create_ca_certificate_role` currently accepts a `certificate` parameter that can either be a PEM-formatted certificate or a path to a PEM-fromatted certificate file. The parameter is differentiated by trying to open a file with what's passed to `certificate`. For some systems, the lengh of a certificate is long enough that it will cause a OSError because of the path length.

The conditional was updated to actually inspect what is passed to contents. To avoid these issues moving forward, `certificate` will only accept a certificate string. Paths should be passed to the new `certificate_file` parameter. For now, passing a certificate file to `certificate` will generate a warning.

Updated tests to use a RSA 4096 key as the certificate file's length would cause the OSError regression solved by this fix. Tests additionally test the certificate and certificate_file parameters.

Signed-off-by: Colin McAllister <colinmca242@gmail.com>